requirements.txt
awscrt~=0.13.5
awsiotsdk~=1.9.3